refactor(nodes): extract text extraction from Element into ElementText
